refactor: add api docs

+++ MisakaBiliApi/Models/ApiResponse/MisakaStreamUrlResponse.cs
@@ -18,8 +18,24 @@ public record MisakaVideoStreamMp4UrlResponse(
     [property: JsonPropertyName("quality")]
     BiliVideoQuality Quality);
 
-public record MisakaVideoStreamDashUrlResponse(MisakaVideoDashItem[] VideoDashs, MisakaVideoDashItem[] AudioDashs, long Duration);
+/// <summary>
+/// DASH 视频流地址响应数据
+/// </summary>
+/// <param name="VideoDashs">视频流 DASH 信息</param>
+/// <param name="AudioDashs">音频流 DASH 信息</param>
+/// <param name="Duration">视频长度，按秒为单位</param>
+public record MisakaVideoStreamDashUrlResponse(MisakaVideoDashItem[] VideoDashs, MisakaVideoDashItem[] AudioDashs, double Duration);
 
+/// <summary>
+/// DASH 流地址信息
+/// </summary>
+/// <param name="Urls">DASH 流的可用 URL</param>
+/// <param name="FrameRate">视频帧率</param>
+/// <param name="Width">视频宽度</param>
+/// <param name="Height">视频高度</param>
+/// <param name="Codecs">编码器</param>
+/// <param name="Bandwidth">码率</param>
+/// <param name="Id">我也不知道这是什么</param>
 public record MisakaVideoDashItem(
     string[] Urls,
     string FrameRate,
